Fork geerlingguy.packer-rhel -> cognifloyd.optimize-vagrantssh
I want to be able to run only the vagrant optimizations when
I'm building a vagrant box, but not at other times.
Therefore, I cut out all of the vm-tools building which
I put in cognifloyd.vm-tools. I use jtyr.vagrant_user to create
the vagrant user. (the fewer the forks, the better).